Use __ptrace_request only on GNU libc
It seems the Posix specification define the ptrace function like this[1]:

    int ptrace(int request, pid_t pid, int addr, int data);

However, the glibc definition is a bit different[2]:

    long ptrace(enum __ptrace_request request, pid_t pid, void *addr, void *data);

Several users has reported compilation errors because of incompatibility
between int and enum. So this patch fix usage of ptrace() for
alternative libc. It seems this is a common way to solve this
problem[3].

[1]: https://www.unix.com/man-page/posix/3c/ptrace/
[2]: https://linux.die.net/man/2/ptrace
[3]: https://github.com/proot-me/proot/blob/master/src/tracee/tracee.h#L38